requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
environs==14.5.0
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.1
magic-filter==1.0.12
marshmallow==4.1.0
multidict==6.7.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
